Consolidate balance calculations
