+++ app/models/account.rb
@@ -1,4 +1,6 @@
 class Account < ApplicationRecord
+  include Syncable
+
   broadcasts_refreshes
   belongs_to :family
   has_many :balances, class_name: "AccountBalance"
@@ -9,19 +11,6 @@ class Account < ApplicationRecord
 
   delegate :type_name, to: :accountable
   before_create :check_currency
-
-  def sync_later
-    AccountSyncJob.perform_later self
-  end
-
-  def sync
-    update!(status: "SYNCING")
-    Account::Syncer.new(self).sync
-    update!(status: "OK")
-  rescue => e
-    update!(status: "ERROR")
-    Rails.logger.error("Failed to sync account #{id}: #{e.message}")
-  end
 
   # Represents the earliest date we can calculate an account balance for
   def effective_start_date
