+++ app/models/account/balance_calculator.rb
@@ -1,0 +1,38 @@
+class Account::BalanceCalculator
+    def initialize(account)
+      @account = account
+    end
+
+    def daily_balances(start_date = nil)
+      sync_start_date = [ start_date, @account.effective_start_date ].compact.max
+
+      valuations = @account.valuations.where("date >= ?", sync_start_date).order(:date).select(:date, :value)
+      transactions = @account.transactions.where("date > ?", sync_start_date).order(:date).select(:date, :amount)
+      oldest_entry = [ valuations.first, transactions.first ].compact.min_by(&:date)
+
+      net_transaction_flows = transactions.sum(&:amount)
+      implied_start_balance = oldest_entry.is_a?(Valuation) ? oldest_entry.value : @account.balance + net_transaction_flows
+
+      prior_balance = implied_start_balance
+      calculated_balances = ((sync_start_date + 1.day)...Date.current).map do |date|
+        valuation = valuations.find { |v| v.date == date }
+
+        if valuation
+            current_balance = valuation.value
+        else
+            current_day_net_transaction_flows = transactions.select { |t| t.date == date }.sum(&:amount)
+            current_balance = prior_balance - current_day_net_transaction_flows
+        end
+
+        prior_balance = current_balance
+
+        { date: date, balance: current_balance, updated_at: Time.current }
+      end
+
+      [
+        { date: sync_start_date, balance: implied_start_balance, updated_at: Time.current },
+        *calculated_balances,
+        { date: Date.current, balance: @account.balance, updated_at: Time.current } # Last balance must always match "source of truth"
+      ]
+    end
+end
